refactored a bit to have only a single function to retreive information from PokeAPI
