+++ basicCommands.go
@@ -1,9 +1,20 @@
 package main
 
 import (
+	"encoding/json"
 	"fmt"
+	"io"
+	"net/http"
 	"os"
+
+	"github.com/marsteg/pokedex/internal/pokecache"
 )
+
+type Pokething interface {
+	Pokemon
+	LocationArea
+	LocationAreas
+}
 
 func help(c *config, _ []string) error {
 	fmt.Printf("Welcome to the Pokedex!\nUsage:\n\n")
@@ -24,4 +35,33 @@ func tiny(c *config, input []string) error {
 	fmt.Println("Tiny is a little lil, but not a Poke-Quil!")
 	fmt.Printf("Command: %v, argument: %v\n", input[0], input[1])
 	return nil
+}
+
+func getPokething[T Pokemon | LocationArea | LocationAreas](url string, c pokecache.Cache) (T, error) {
+	var pokemon T
+
+	val, exists := c.Get(url)
+	if !exists {
+		res, err := http.Get(url)
+		if err != nil {
+			fmt.Println(err)
+		}
+		body, err := io.ReadAll(res.Body)
+		if err != nil {
+			return pokemon, err
+		}
+		res.Body.Close()
+		c.Add(url, body)
+		err = json.Unmarshal(body, &pokemon)
+		if err != nil {
+			return pokemon, err
+		}
+		return pokemon, nil
+	} else {
+		err := json.Unmarshal(val, &pokemon)
+		if err != nil {
+			return pokemon, err
+		}
+		return pokemon, nil
+	}
 }
