+++ catch.go
@@ -1,293 +1,17 @@
 package main
 
 import (
-	"encoding/json"
 	"errors"
 	"fmt"
-	"io"
 	"math/rand"
-	"net/http"
 	"time"
-
-	"github.com/marsteg/pokedex/internal/pokecache"
 )
-
-type Pokemon struct {
-	Abilities []struct {
-		Ability struct {
-			Name string `json:"name"`
-			URL  string `json:"url"`
-		} `json:"ability"`
-		IsHidden bool `json:"is_hidden"`
-		Slot     int  `json:"slot"`
-	} `json:"abilities"`
-	BaseExperience int `json:"base_experience"`
-	Cries          struct {
-		Latest string `json:"latest"`
-		Legacy string `json:"legacy"`
-	} `json:"cries"`
-	Forms []struct {
-		Name string `json:"name"`
-		URL  string `json:"url"`
-	} `json:"forms"`
-	GameIndices []struct {
-		GameIndex int `json:"game_index"`
-		Version   struct {
-			Name string `json:"name"`
-			URL  string `json:"url"`
-		} `json:"version"`
-	} `json:"game_indices"`
-	Height    int `json:"height"`
-	HeldItems []struct {
-		Item struct {
-			Name string `json:"name"`
-			URL  string `json:"url"`
-		} `json:"item"`
-		VersionDetails []struct {
-			Rarity  int `json:"rarity"`
-			Version struct {
-				Name string `json:"name"`
-				URL  string `json:"url"`
-			} `json:"version"`
-		} `json:"version_details"`
-	} `json:"held_items"`
-	ID                     int    `json:"id"`
-	IsDefault              bool   `json:"is_default"`
-	LocationAreaEncounters string `json:"location_area_encounters"`
-	Moves                  []struct {
-		Move struct {
-			Name string `json:"name"`
-			URL  string `json:"url"`
-		} `json:"move"`
-		VersionGroupDetails []struct {
-			LevelLearnedAt  int `json:"level_learned_at"`
-			MoveLearnMethod struct {
-				Name string `json:"name"`
-				URL  string `json:"url"`
-			} `json:"move_learn_method"`
-			VersionGroup struct {
-				Name string `json:"name"`
-				URL  string `json:"url"`
-			} `json:"version_group"`
-		} `json:"version_group_details"`
-	} `json:"moves"`
-	Name          string `json:"name"`
-	Order         int    `json:"order"`
-	PastAbilities []any  `json:"past_abilities"`
-	PastTypes     []any  `json:"past_types"`
-	Species       struct {
-		Name string `json:"name"`
-		URL  string `json:"url"`
-	} `json:"species"`
-	Sprites struct {
-		BackDefault      string `json:"back_default"`
-		BackFemale       string `json:"back_female"`
-		BackShiny        string `json:"back_shiny"`
-		BackShinyFemale  string `json:"back_shiny_female"`
-		FrontDefault     string `json:"front_default"`
-		FrontFemale      string `json:"front_female"`
-		FrontShiny       string `json:"front_shiny"`
-		FrontShinyFemale string `json:"front_shiny_female"`
-		Other            struct {
-			DreamWorld struct {
-				FrontDefault string `json:"front_default"`
-				FrontFemale  any    `json:"front_female"`
-			} `json:"dream_world"`
-			Home struct {
-				FrontDefault     string `json:"front_default"`
-				FrontFemale      string `json:"front_female"`
-				FrontShiny       string `json:"front_shiny"`
-				FrontShinyFemale string `json:"front_shiny_female"`
-			} `json:"home"`
-			OfficialArtwork struct {
-				FrontDefault string `json:"front_default"`
-				FrontShiny   string `json:"front_shiny"`
-			} `json:"official-artwork"`
-			Showdown struct {
-				BackDefault      string `json:"back_default"`
-				BackFemale       string `json:"back_female"`
-				BackShiny        string `json:"back_shiny"`
-				BackShinyFemale  any    `json:"back_shiny_female"`
-				FrontDefault     string `json:"front_default"`
-				FrontFemale      string `json:"front_female"`
-				FrontShiny       string `json:"front_shiny"`
-				FrontShinyFemale string `json:"front_shiny_female"`
-			} `json:"showdown"`
-		} `json:"other"`
-		Versions struct {
-			GenerationI struct {
-				RedBlue struct {
-					BackDefault      string `json:"back_default"`
-					BackGray         string `json:"back_gray"`
-					BackTransparent  string `json:"back_transparent"`
-					FrontDefault     string `json:"front_default"`
-					FrontGray        string `json:"front_gray"`
-					FrontTransparent string `json:"front_transparent"`
-				} `json:"red-blue"`
-				Yellow struct {
-					BackDefault      string `json:"back_default"`
-					BackGray         string `json:"back_gray"`
-					BackTransparent  string `json:"back_transparent"`
-					FrontDefault     string `json:"front_default"`
-					FrontGray        string `json:"front_gray"`
-					FrontTransparent string `json:"front_transparent"`
-				} `json:"yellow"`
-			} `json:"generation-i"`
-			GenerationIi struct {
-				Crystal struct {
-					BackDefault           string `json:"back_default"`
-					BackShiny             string `json:"back_shiny"`
-					BackShinyTransparent  string `json:"back_shiny_transparent"`
-					BackTransparent       string `json:"back_transparent"`
-					FrontDefault          string `json:"front_default"`
-					FrontShiny            string `json:"front_shiny"`
-					FrontShinyTransparent string `json:"front_shiny_transparent"`
-					FrontTransparent      string `json:"front_transparent"`
-				} `json:"crystal"`
-				Gold struct {
-					BackDefault      string `json:"back_default"`
-					BackShiny        string `json:"back_shiny"`
-					FrontDefault     string `json:"front_default"`
-					FrontShiny       string `json:"front_shiny"`
-					FrontTransparent string `json:"front_transparent"`
-				} `json:"gold"`
-				Silver struct {
-					BackDefault      string `json:"back_default"`
-					BackShiny        string `json:"back_shiny"`
-					FrontDefault     string `json:"front_default"`
-					FrontShiny       string `json:"front_shiny"`
-					FrontTransparent string `json:"front_transparent"`
-				} `json:"silver"`
-			} `json:"generation-ii"`
-			GenerationIii struct {
-				Emerald struct {
-					FrontDefault string `json:"front_default"`
-					FrontShiny   string `json:"front_shiny"`
-				} `json:"emerald"`
-				FireredLeafgreen struct {
-					BackDefault  string `json:"back_default"`
-					BackShiny    string `json:"back_shiny"`
-					FrontDefault string `json:"front_default"`
-					FrontShiny   string `json:"front_shiny"`
-				} `json:"firered-leafgreen"`
-				RubySapphire struct {
-					BackDefault  string `json:"back_default"`
-					BackShiny    string `json:"back_shiny"`
-					FrontDefault string `json:"front_default"`
-					FrontShiny   string `json:"front_shiny"`
-				} `json:"ruby-sapphire"`
-			} `json:"generation-iii"`
-			GenerationIv struct {
-				DiamondPearl struct {
-					BackDefault      string `json:"back_default"`
-					BackFemale       string `json:"back_female"`
-					BackShiny        string `json:"back_shiny"`
-					BackShinyFemale  string `json:"back_shiny_female"`
-					FrontDefault     string `json:"front_default"`
-					FrontFemale      string `json:"front_female"`
-					FrontShiny       string `json:"front_shiny"`
-					FrontShinyFemale string `json:"front_shiny_female"`
-				} `json:"diamond-pearl"`
-				HeartgoldSoulsilver struct {
-					BackDefault      string `json:"back_default"`
-					BackFemale       string `json:"back_female"`
-					BackShiny        string `json:"back_shiny"`
-					BackShinyFemale  string `json:"back_shiny_female"`
-					FrontDefault     string `json:"front_default"`
-					FrontFemale      string `json:"front_female"`
-					FrontShiny       string `json:"front_shiny"`
-					FrontShinyFemale string `json:"front_shiny_female"`
-				} `json:"heartgold-soulsilver"`
-				Platinum struct {
-					BackDefault      string `json:"back_default"`
-					BackFemale       string `json:"back_female"`
-					BackShiny        string `json:"back_shiny"`
-					BackShinyFemale  string `json:"back_shiny_female"`
-					FrontDefault     string `json:"front_default"`
-					FrontFemale      string `json:"front_female"`
-					FrontShiny       string `json:"front_shiny"`
-					FrontShinyFemale string `json:"front_shiny_female"`
-				} `json:"platinum"`
-			} `json:"generation-iv"`
-			GenerationV struct {
-				BlackWhite struct {
-					Animated struct {
-						BackDefault      string `json:"back_default"`
-						BackFemale       string `json:"back_female"`
-						BackShiny        string `json:"back_shiny"`
-						BackShinyFemale  string `json:"back_shiny_female"`
-						FrontDefault     string `json:"front_default"`
-						FrontFemale      string `json:"front_female"`
-						FrontShiny       string `json:"front_shiny"`
-						FrontShinyFemale string `json:"front_shiny_female"`
-					} `json:"animated"`
-					BackDefault      string `json:"back_default"`
-					BackFemale       string `json:"back_female"`
-					BackShiny        string `json:"back_shiny"`
-					BackShinyFemale  string `json:"back_shiny_female"`
-					FrontDefault     string `json:"front_default"`
-					FrontFemale      string `json:"front_female"`
-					FrontShiny       string `json:"front_shiny"`
-					FrontShinyFemale string `json:"front_shiny_female"`
-				} `json:"black-white"`
-			} `json:"generation-v"`
-			GenerationVi struct {
-				OmegarubyAlphasapphire struct {
-					FrontDefault     string `json:"front_default"`
-					FrontFemale      string `json:"front_female"`
-					FrontShiny       string `json:"front_shiny"`
-					FrontShinyFemale string `json:"front_shiny_female"`
-				} `json:"omegaruby-alphasapphire"`
-				XY struct {
-					FrontDefault     string `json:"front_default"`
-					FrontFemale      string `json:"front_female"`
-					FrontShiny       string `json:"front_shiny"`
-					FrontShinyFemale string `json:"front_shiny_female"`
-				} `json:"x-y"`
-			} `json:"generation-vi"`
-			GenerationVii struct {
-				Icons struct {
-					FrontDefault string `json:"front_default"`
-					FrontFemale  any    `json:"front_female"`
-				} `json:"icons"`
-				UltraSunUltraMoon struct {
-					FrontDefault     string `json:"front_default"`
-					FrontFemale      string `json:"front_female"`
-					FrontShiny       string `json:"front_shiny"`
-					FrontShinyFemale string `json:"front_shiny_female"`
-				} `json:"ultra-sun-ultra-moon"`
-			} `json:"generation-vii"`
-			GenerationViii struct {
-				Icons struct {
-					FrontDefault string `json:"front_default"`
-					FrontFemale  string `json:"front_female"`
-				} `json:"icons"`
-			} `json:"generation-viii"`
-		} `json:"versions"`
-	} `json:"sprites"`
-	Stats []struct {
-		BaseStat int `json:"base_stat"`
-		Effort   int `json:"effort"`
-		Stat     struct {
-			Name string `json:"name"`
-			URL  string `json:"url"`
-		} `json:"stat"`
-	} `json:"stats"`
-	Types []struct {
-		Slot int `json:"slot"`
-		Type struct {
-			Name string `json:"name"`
-			URL  string `json:"url"`
-		} `json:"type"`
-	} `json:"types"`
-	Weight int `json:"weight"`
-}
 
 func catch(c *config, input []string) error {
 	pokemon := input[1]
 
 	url := "https://pokeapi.co/api/v2/pokemon/" + pokemon + "/"
-	reply, err := getPokemon(url, c.cache)
+	reply, err := getPokething[Pokemon](url, c.cache)
 	if err != nil {
 		if err.Error() == "invalid character 'N' looking for beginning of value" {
 			fmt.Printf("failed throwing a Pokeball at " + pokemon + "!!\n")
@@ -315,32 +39,3 @@ func catch(c *config, input []string) error {
 //Once the Pokemon is caught, add it to the user's Pokedex. I used a map[string]Pokemon
 // to keep track of caught Pokemon.
 //You'll want to store the Pokemon's data so that in the next step we can use it. (caching)
-
-func getPokemon(url string, c pokecache.Cache) (Pokemon, error) {
-	var pokemon Pokemon
-
-	val, exists := c.Get(url)
-	if !exists {
-		res, err := http.Get(url)
-		if err != nil {
-			fmt.Println(err)
-		}
-		body, err := io.ReadAll(res.Body)
-		if err != nil {
-			return pokemon, err
-		}
-		res.Body.Close()
-		c.Add(url, body)
-		err = json.Unmarshal(body, &pokemon)
-		if err != nil {
-			return pokemon, err
-		}
-		return pokemon, nil
-	} else {
-		err := json.Unmarshal(val, &pokemon)
-		if err != nil {
-			return pokemon, err
-		}
-		return pokemon, nil
-	}
-}
