+++ map.go
@@ -9,19 +9,9 @@ import (
 	"github.com/marsteg/pokedex/internal/pokecache"
 )
 
-type response struct {
-	Count    int    `json:"count"`
-	Next     string `json:"next"`
-	Previous string `json:"previous,omitempty"`
-	Results  []struct {
-		Name string `json:"name"`
-		URL  string `json:"url"`
-	} `json:"results"`
-}
-
 func commandMap(c *config, _ []string) error {
 	if c.next == "" {
-		reply, err := getreply("https://pokeapi.co/api/v2/location-area/", c.cache)
+		reply, err := getPokething[LocationAreas]("https://pokeapi.co/api/v2/location-area/", c.cache)
 		if err != nil {
 			fmt.Println(err)
 		}
@@ -51,7 +41,7 @@ func commandMapB(c *config, _ []string) error {
 		fmt.Println("error, go forward on the map first")
 		return nil
 	} else {
-		reply, err := getreply(c.previous, c.cache)
+		reply, err := getPokething[LocationAreas](c.previous, c.cache)
 		if err != nil {
 			fmt.Println(err)
 		}
@@ -65,12 +55,8 @@ func commandMapB(c *config, _ []string) error {
 	}
 }
 
-func getreply(url string, c pokecache.Cache) (response, error) {
-	var reply response
-
-	//If you already have the data for a given URL (which is our cache key)
-	//in the cache, you should use that instead of making a new request.
-	//Whenever you do make a request, you should add the response to the cache.
+func getreply(url string, c pokecache.Cache) (LocationAreas, error) {
+	var reply LocationAreas
 	val, exists := c.Get(url)
 	if !exists {
 		res, err := http.Get(url)
